fix: move types to separate file for menu items
